test commit

+++ ProsperLend/src/main/java/com/model/service/LoanService.java
@@ -1,6 +1,5 @@
 package com.model.service;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 import com.dto.entity.Loans;
@@ -13,7 +12,7 @@ public interface LoanService {
 
 	boolean addLoan(Loans loan);
 
-	boolean changeAmountInterest(int loanId, BigDecimal newAmount, BigDecimal newInterest);
+	boolean changeAmountInterest(int loanId, Double newAmount, Double newInterest);
